refactor jinja utils, extracting from openshift_resources_base
